- Fixed unnecessary resource allocation/deallocation in Socket. - Added Process() function in Transporter and Physical which allows processing of additional tasks other than Send and Receive. TCP Server uses Process() to Accept client connections - Reflected these changes in TransportApp

+++ Transport/Message.cpp
@@ -65,7 +65,7 @@ namespace Transport
 
 		for (DataType::Size i = 0; i < sizeof(value); i++)
 		{
-			bytes.push_back(value >> (i * 8));
+			bytes.push_back(static_cast<DataType::Byte>(value >> (i * 8)));
 		}
 
 		data.insert(data.end(), bytes.begin(), bytes.end());
